requirements.txt
Django~=5.2.4
djangorestframework~=3.16.0
django-smart-selects~=1.7.2
dotenv~=0.9.9
python-dotenv~=1.1.1